Added conformance tester https://github.com/knative-sandbox/control-protocol/pull/58

+++ data-plane/control-protocol/src/main/java/dev/knative/eventing/control/protocol/ControlMessageHeader.java
@@ -22,7 +22,7 @@ public interface ControlMessageHeader {
   /**
    * @return the message op code
    */
-  short opCode();
+  byte opCode();
 
   /**
    * @return the message UUID
