+++ data-plane/control-protocol/src/main/java/dev/knative/eventing/control/protocol/impl/ControlMessageImpl.java
@@ -100,7 +100,7 @@ public class ControlMessageImpl implements ControlMessage {
   }
 
   @Override
-  public short opCode() {
+  public byte opCode() {
     return this.opCode;
   }
